\name{plot.dr}
\alias{plot}
\alias{plot.dr}
\title{Generic method for plotting 'dr' class objects.}
\description{Generic method for plotting 'dr' class objects.}
\usage{\method{plot}{dr}(x, ...)}
\arguments{
  \item{x}{dr object. This is the output from \code{dead_reckoning}}
  \item{...}{ignored}
  }
\seealso{
  \code{\link{dead_reckoning}}.
}


