\name{print.dr}
\alias{print}
\alias{print.dr}
\title{Generic method for printing information about 'dr' class objects.}
\description{Generic method for printing information about 'dr' class objects.}
\usage{\method{print}{dr}(x, ...)}
\arguments{
  \item{x}{dr object. This is the output from \code{dead_reckoning}}
  \item{...}{ignored}
  }
\seealso{
  \code{\link{dead_reckoning}}.
}


