\name{utmzones}
\alias{utmzones}
\docType{data}
\title{utmzone data}
\description{utmzone data spatial polygons data frame}
\usage{data(utmzones)}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\source{National Geospatial-Intelligence Agency, \url{http://earth-info.nga.mil/GandG/coordsys/grids/universal_grid_system.html}
}
\examples{
data(utmzones)
}
\keyword{datasets}
